// web/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ProblemDetail: {
            /** Type */
            type: string;
            /** Title */
            title: string;
            /** Status */
            status: number;
            /** Detail */
            detail?: string | null;
            /** Instance */
            instance?: string | null;
            /** Errors */
            errors?: components["schemas"]["FieldError"][] | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}